refactor: tratar aluno entity

+++ backend/src/main/java/com/denisson/backend/domain/useCases/alunos/DeleteAlunoUseCase.java
@@ -1,5 +1,7 @@
 package com.denisson.backend.domain.useCases.alunos;
 
+import com.denisson.backend.domain.exceptions.NotFoundException;
+import com.denisson.backend.domain.models.Aluno;
 import com.denisson.backend.domain.repositories.IAlunoRepository;
 
 public class DeleteAlunoUseCase {
@@ -10,6 +12,10 @@ public class DeleteAlunoUseCase {
     }
 
     public void execute(Long id) {
+        Aluno existsAluno = repository.findById(id);
+        if (existsAluno == null) {
+            throw new NotFoundException("Aluno não encontrado com id " + id);
+        }
         repository.delete(id);
     }
 }
